lint: clear srt-transport's clippy -D warnings debt
Eight lints, all in srt-transport, each fixed rather than allowed:

- unnecessary_cast (x5, epoll_bridge): the event word is now read as
  `SRT_EPOLL_OPT` instead of `u32`, so the flag constants are tested with
  no cast at all. C leaves an enum's underlying type to the compiler:
  bindgen emits `c_uint` for it under GCC/Clang (x86_64 and aarch64
  alike) but `c_int` under MSVC, where the old `as u32` on each constant
  was load-bearing. Casting to the alias keeps every target correct
  without a per-target allow.
- unnecessary_cast (stats_bridge): `pktCongestionWindow` is already
  `c_int` = `i32`, the type of `SrtStats::pkt_congestion_window`.
- unnecessary_mut_passed: `srt_epoll_add_usock` takes `const int*`, so
  `add_to_epoll` passes `&events` and drops its mutable copy.
- unnecessary_map_or: `map_or(true, ..)` -> `is_none_or(..)`.

Behaviour is unchanged; no public signature is touched (all of it is
private to the crate).

+++ srt-transport/src/stats_bridge.rs
@@ -75,7 +75,7 @@ pub(crate) fn convert_perfmon_to_stats(perf: &CBytePerfMon) -> SrtStats {
         // Instant measurements
         us_pkt_snd_period: perf.usPktSndPeriod,
         pkt_flow_window: perf.pktFlowWindow,
-        pkt_congestion_window: perf.pktCongestionWindow as i32,
+        pkt_congestion_window: perf.pktCongestionWindow,
         pkt_flight_size: perf.pktFlightSize,
         ms_rtt: perf.msRTT,
         mbps_bandwidth: perf.mbpsBandwidth,
